Allow building as static library
i'm working in an embedded context without the luxury of dyanmic
libraries and linking. removing the explicit `SHARED` from the
add_library call apparently allows both shared and static builds.
the target_include_directories call is necessary to provide "cat.h"
to the dependant when the library is linked by cmake.

-barrow